Move integration test method generatePeers() from sync to network setup
