+++ test/integration/setup/index.js
@@ -19,7 +19,6 @@ var utils = require('../utils');
 var network = require('./network');
 var pm2 = require('./pm2');
 var shell = require('./shell');
-var sync = require('./sync');
 
 var WAIT_BEFORE_CONNECT_MS = 20000;
 
@@ -71,5 +70,4 @@ module.exports = {
 	network,
 	pm2,
 	shell,
-	sync,
 };
