+++ test/integration/setup/network.js
@@ -20,6 +20,21 @@ var waitUntilBlockchainReady = require('../../common/utils/wait_for')
 	.blockchainReady;
 var utils = require('../utils');
 
+var SYNC_MODES = {
+	RANDOM: 0,
+	ALL_TO_FIRST: 1,
+	ALL_TO_GROUP: 2,
+};
+
+var SYNC_MODE_DEFAULT_ARGS = {
+	RANDOM: {
+		probability: 0.5, // (0 - 1)
+	},
+	ALL_TO_GROUP: {
+		indices: [],
+	},
+};
+
 module.exports = {
 	waitForAllNodesToBeReady(configurations, cb) {
 		const retries = 20;
@@ -65,4 +80,64 @@ module.exports = {
 				return cb(error);
 			});
 	},
+
+	generatePeers(configurations, syncMode, syncModeArgs, currentPeer) {
+		syncModeArgs = syncModeArgs || SYNC_MODE_DEFAULT_ARGS[syncMode];
+
+		var peersList = [];
+
+		switch (syncMode) {
+			case SYNC_MODES.RANDOM:
+				if (typeof syncModeArgs.probability !== 'number') {
+					throw new Error(
+						'Probability parameter not specified to random sync mode'
+					);
+				}
+				var isPickedWithProbability = function(n) {
+					return !!n && Math.random() <= n;
+				};
+				configurations.forEach(configuration => {
+					if (isPickedWithProbability(syncModeArgs.probability)) {
+						if (!(configuration.wsPort === currentPeer)) {
+							peersList.push({
+								ip: configuration.ip,
+								wsPort: configuration.wsPort,
+							});
+						}
+					}
+				});
+				break;
+
+			case SYNC_MODES.ALL_TO_FIRST:
+				if (configurations.length === 0) {
+					throw new Error('No configurations provided');
+				}
+				peersList = [
+					{
+						ip: configurations[0].ip,
+						wsPort: configurations[0].wsPort,
+					},
+				];
+				break;
+
+			case SYNC_MODES.ALL_TO_GROUP:
+				if (!Array.isArray(syncModeArgs.indices)) {
+					throw new Error('Provide peers indices to sync with as an array');
+				}
+				configurations.forEach((configuration, index) => {
+					if (syncModeArgs.indices.indexOf(index) !== -1) {
+						if (!(configuration.wsPort === currentPeer)) {
+							peersList.push({
+								ip: configuration.ip,
+								wsPort: configuration.wsPort,
+							});
+						}
+					}
+				});
+			// no default
+		}
+
+		return peersList;
+	},
+	SYNC_MODES,
 };
